Move descriptor builder
