+++ src/di/descriptor.rs
@@ -1,7 +1,6 @@
 use crate::{ServiceDependency, ServiceProvider, Type};
 use spin::Once;
 use std::any::Any;
-use std::marker::PhantomData;
 
 /// Represents the possible service lifetimes.
 #[derive(Copy, Clone, Debug, PartialEq)]
@@ -118,84 +117,5 @@ impl Clone for ServiceDescriptor {
         // without context, we don't know if this is 'safe';
         // always copy dependencies here
         self.clone_with(true)
-    }
-}
-
-/// Represents a builder for [service descriptors](struct.ServiceDescriptor.html).
-pub struct ServiceDescriptorBuilder<TSvc: Any + ?Sized, TImpl> {
-    lifetime: ServiceLifetime,
-    service_type: Type,
-    implementation_type: Type,
-    dependencies: Vec<ServiceDependency>,
-    _marker_svc: PhantomData<TSvc>,
-    _marker_impl: PhantomData<TImpl>,
-}
-
-impl<TSvc: Any + ?Sized, TImpl> ServiceDescriptorBuilder<TSvc, TImpl> {
-    /// Defines the factory method used to activate the service and returns the service descriptor.
-    ///
-    /// # Arguments
-    ///
-    /// * `factory` - The factory method used to create the service
-    pub fn from(mut self, factory: fn(&ServiceProvider) -> ServiceRef<TSvc>) -> ServiceDescriptor {
-        ServiceDescriptor {
-            lifetime: self.lifetime,
-            service_type: self.service_type,
-            implementation_type: self.implementation_type,
-            dependencies: if self.dependencies.is_empty() {
-                Vec::with_capacity(0)
-            } else {
-                self.dependencies.shrink_to_fit();
-                self.dependencies
-            },
-            instance: ServiceRef::new(Once::new()),
-            factory: ServiceRef::new(move |sp| ServiceRef::new(factory(sp))),
-        }
-    }
-
-    /// Defines a dependency used by the service.
-    ///
-    /// # Arguments
-    ///
-    /// * `dependency` - The [dependency](struct.ServiceDependency.html) associated with the service
-    pub fn depends_on(mut self, dependency: ServiceDependency) -> Self {
-        if !self.dependencies.contains(&dependency) {
-            self.dependencies.push(dependency);
-        }
-        self
-    }
-
-    /// Initializes a new service descriptor builder.
-    ///
-    /// # Arguments
-    ///
-    /// * `lifetime` - The [lifetime](enum.ServiceLifetime.html) of the service
-    /// * `implementation_type` - The service implementation [type](struct.Type.html)
-    pub fn new(lifetime: ServiceLifetime, implementation_type: Type) -> Self {
-        Self {
-            lifetime,
-            service_type: Type::of::<TSvc>(),
-            implementation_type,
-            dependencies: Vec::new(),
-            _marker_svc: PhantomData,
-            _marker_impl: PhantomData,
-        }
-    }
-
-    /// Initializes a new service descriptor builder.
-    ///
-    /// # Arguments
-    ///
-    /// * `lifetime` - The [lifetime](enum.ServiceLifetime.html) of the service
-    /// * `implementation_type` - The service implementation [type](struct.Type.html)
-    pub fn keyed<TKey>(lifetime: ServiceLifetime, implementation_type: Type) -> Self {
-        Self {
-            lifetime,
-            service_type: Type::keyed::<TKey, TSvc>(),
-            implementation_type,
-            dependencies: Vec::new(),
-            _marker_svc: PhantomData,
-            _marker_impl: PhantomData,
-        }
     }
 }
